requirements.txt
langchain-openai
langchain
pydantic
python-dotenv
